封装分组相关的前端API和页面

+++ hopesfire-dao/src/main/java/com/kxw/hopesfire/dao/entity/GroupEntity.java
@@ -16,7 +16,7 @@ import lombok.EqualsAndHashCode;
 public class GroupEntity extends BaseEntity {
 
     private String name;
-    private int tenantId;
+    private Long tenantId;
     private String remark;
 
 }
